+++ src/controllers/NguoiDung/deleteNguoiDung.js
@@ -4,15 +4,26 @@ const { sucessCode, failCode, errorCode } = require("../../config/reponse");
 const deleteNguoiDung = async (req, res) => {
     let id = req.params.id;
     try {
-        const data = await prisma.nguoiDung.delete({
+        let checkNguoiDung = await prisma.nguoiDung.findFirst({
             where: {
-                id: Number(id),
-            },
+                id: Number(id)
+            }
         });
-        if (data == "") {
+        if (!checkNguoiDung) {
             failCode(res, "", "Xóa người dùng thất bại !");
-        } else {
-            sucessCode(res, data, "Xóa người dùng thành công !");
+
+        }
+        else {
+            const data = await prisma.nguoiDung.delete({
+                where: {
+                    id: Number(id),
+                },
+            });
+            if (data == "") {
+                failCode(res, "", "Xóa người dùng thất bại !");
+            } else {
+                sucessCode(res, data, "Xóa người dùng thành công !");
+            }
         }
     } catch (error) {
         errorCode(res, "Lỗi Backend");
